Shift loading test over to the demo app.
