+++ SukiUI.Demo/SukiUI.Demo.csproj
@@ -57,6 +57,8 @@
 	</ItemGroup>
 
 	<ItemGroup>
+	  <EmbeddedResource Include="Assets\glow.sksl" />
+	  <EmbeddedResource Include="Assets\pellets.sksl" />
 	  <EmbeddedResource Include="Assets\space.sksl" />
 	  <EmbeddedResource Include="Assets\clouds.sksl" />
 	  <EmbeddedResource Include="Assets\weird.sksl" />
